requirements.txt
openai
python-dotenv
SpeechRecognition
pyaudio
pynput
gTTS
pydub